learn . net core

+++ Models/Products.cs
@@ -1,24 +1,25 @@
-// using System;
+ using System;
 // using Microsoft.AspNetCore.Mvc; 
+using System.ComponentModel.DataAnnotations;
 namespace nwl_dotnetcore.Models
 {
     public class Products
     {
-        public int ID {get; set;}
-        public string Name{get;set;}
+        public int ID { get; set; }
+        public string Name { get; set; }
         // public string ProductId{get;set;}
-        public string Detail{get;set;}
-        public decimal Price{get;set;}
+        public string Detail { get; set; }
+        public int Price { get; set; }
         // public int CategoryId{get;set;}
         // public string Image{get;set;}
         // public DateTime Timestamp{get;set;}
         // public virtual Category Category{get;set;}
-        
-        
-        
-        
-        
-        
-        
+
+
+
+
+
+
+
     }
 }
